handle touch events directly in AbsRecyclerViewFastScroller
also remove class FastScrollerTouchListene
This change will allow user to set custom OnTouchListener without redefining default one
